+++ noodles-bcf/src/header/string_maps.rs
@@ -43,7 +43,7 @@ impl StringMaps {
     /// let header = vcf::Header::builder()
     ///     .add_info(info::Key::TotalDepth, Map::<Info>::from(&info::Key::TotalDepth))
     ///     .add_filter("q10", Map::<Filter>::new("Quality below 10"))
-    ///     .add_format(format::Key::ReadDepth, Map::<Format>::from(format::Key::ReadDepth))
+    ///     .add_format(format::Key::ReadDepth, Map::<Format>::from(&format::Key::ReadDepth))
     ///     .add_contig("sq0".parse()?, Map::<Contig>::new())
     ///     .build();
     ///
@@ -78,7 +78,7 @@ impl StringMaps {
     /// let header = vcf::Header::builder()
     ///     .add_info(info::Key::TotalDepth, Map::<Info>::from(&info::Key::TotalDepth))
     ///     .add_filter("q10", Map::<Filter>::new("Quality below 10"))
-    ///     .add_format(format::Key::ReadDepth, Map::<Format>::from(format::Key::ReadDepth))
+    ///     .add_format(format::Key::ReadDepth, Map::<Format>::from(&format::Key::ReadDepth))
     ///     .add_contig("sq0".parse()?, Map::<Contig>::new())
     ///     .build();
     ///
@@ -417,11 +417,11 @@ mod tests {
             .add_filter("q10", Map::<Filter>::new("Quality below 10"))
             .add_format(
                 format::Key::Genotype,
-                Map::<Format>::from(format::Key::Genotype),
+                Map::<Format>::from(&format::Key::Genotype),
             )
             .add_format(
                 format::Key::ReadDepth,
-                Map::<Format>::from(format::Key::ReadDepth),
+                Map::<Format>::from(&format::Key::ReadDepth),
             )
             .add_alternative_allele(del, Map::<AlternativeAllele>::new("Deletion"))
             .build();